Library/Camera: Implement `ActorCameraSubTarget`
